Changing name and return type of getTransaction
getTransaction implies that you get a transaction. However only a
pointer memory with the size of a transaction is returned. By renaming
it to allocateTransaction the name better represents its functionality.

The return type of the function was a `std::optional` because for most
types an state where a value is not available is not defined. For a raw
pointer however an nullptr can be used. Because of this the return type
has been changed to a raw pointer. This also gets rid of the `.value()`
everytime the transaction is accesed.