Only statically allocate in display driver and remove lvlg from menuconfig. (#315)
* merge

* fixing separate dma writes for pos set

* The setPos function of the display isn't allocating anymore

* tft035: adding a retry for the setpos
If the setpos fails a retry will take place now to ensure that the
screen is drawn correctly. Their should also be a log but for that a
good logging system should be in place.

* adding logging for the setpos retry

* moving lvgl out of components manually adding include folders and srcs

* cleaning up the cmake file

* fixing error in pushing wrong compiled proto

* resetting protobuf files hard to develop

* fixing protobuf

* removing unnecessary include in cmakefile for lvgl

* correcting spelling error in docs

* removing unnecessary casts

* adding wait and timeout to the retry for getting a transaction from the
transaction buffer

* changing writing spi dma as value to be passed by array instead of
uint32_t

* changing spi timeout for getting transaction from buffer to 10x 1ms

* only allowing vectors of uint8_t for value based spi transactions

* removing unnecessary reinterpret cast

* fixing wrong docs

* fixing wrong use of const

* adding the getTransaction function in the spi driver to remove
duplicate code

* fixing const issue

* Changing name and return type of getTransaction

getTransaction implies that you get a transaction. However only a
pointer memory with the size of a transaction is returned. By renaming
it to allocateTransaction the name better represents its functionality.

The return type of the function was a `std::optional` because for most
types an state where a value is not available is not defined. For a raw
pointer however an nullptr can be used. Because of this the return type
has been changed to a raw pointer. This also gets rid of the `.value()`
everytime the transaction is accesed.

* improve docs of the spi library

* improve readabilty of allocatetransaction function

* split data en command functions and harmonise naming